requirements-deploy.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
dnspython==2.4.2
openai==1.109.1
pydantic==2.5.0
aiofiles==23.2.1